abstract class tuner created
